universal-ckp: reconfigure model parameter rng tracker
When loading from universal checkpoint with a different model parameter
configuration, the loaded tensor parallel RNG tracker states are incorrect.
In this case, we reconfigure the tensor parallel RNG tracker states with new
seed values (each tp rank with a unique seed).
We add an offset=iteration to the base seed. This is to ensure that when we
load multiple times from universal checkpoint, we will use a different random
sequence at each run.

This commit requires a counter change in DeepSpeed repo.

Signed-off-by: Moshe Island <misland@habana.ai>